Among other things, an attent to rename new functions to follow the SYSTEM_actionItem naming convention loosely in place in the VCU
Enabling a different way to make a torque request during power limiting. Enabling a positive feedback loop between pl and the mcm output so long as it is less than the driver requested torque

Adding 80KWH setpoint to plPID initialization

removing functions that are not used in plPID method

commenting out can messages that im not using

Another one of my function&variable name formatting commits. While it may seem like I enjoy doing this, I don't. A naming convention should be established and applied to all variable & function names. While not a big deal in the context of making a car run, it is also simultaneously extremely important, as poor/inconsistant naming conventions is still one of the biggest barriers to understanding what is happening in our code at any given point, especially from the perspective of a first time viewer. This problem is exacerbated when important information such as 'from what struct are we getting a value?' is left out of the name. Such issues needlessly worsen the learning curve, and is a barrier that can be easily removed by taking a few extra seconds and being proper in all areas of coding, not just ones we deem worthy based upon a whim. These decisions should not be made unilaterally by me. A unified decision should be made by all involved parties. -rant done