integration w other subsystems(PL>LC)
WIP for powerLimit LUT. I am aware of the fact that typings are not consistent in calculations and other various programming errors. The intention of this commit is to allow the current logic to be viewable by others and I plan to fix the issues (and do an overhaul of a number of things, including merging plpPID back into this branch, within the next week or two. Also upcoming if I can manage, #define statements will be used to switch between states, and to enable/disable verbose logging over CAN.

Partial update. Fixed an issue with the math for double interpolation, and slowly using better variable types.