Code cleanup, lots of float4s changes to s/ubytes of various sizes because the casting was not needed. Other changes as well.
Finally a working ifdef solution. Also, changed some things in Power PID to be casted as float4 since the integer math would just produce 0, instead of the intended values.

Major updates to LUT MEthod of Powerlimiting. Need to remember that there are n-1 steps in a n sized array lmao. Also, I forgot to factor in the difference between the lower bounds of the LUT and 0, because 2000 % 160 /= 0, which was messing up a lot of calculations. Did some float4 casting to preserve decimal multiplication.

Fixing the hashtable insert pair function

the values were not sent into the table. the HashTable_insertPair() function took in the wrong parameter, fixed now to input the hashtable as a param.

Fix to everything.

Issues from Rebase affecting first build. Still need to review diff to make sure I didn't mess anything up of our code