Extremely Large Rework to PL code. Switching from using float4 to ubyte/sbyte and using deci- units for both PID internal math & Powerlimit code. The PID struct has now been moved back into the PL struct, so no more separation of the two. Additionally, the DAQ repo is now submoduled in order to quickly reference the DBC. Furthermore, PL code now supports multiple different PL HashTables in preparation for rotary switches on SR-16 for active derating. This code needs to be tested in order to confirm a working car and consistency with previous behavior of PL. Some of the new structure of PL (the PID as part of the struct and the use of deci-Nm over float4) will be added to Launch Controlcode in the future. Additionally PL does not currently support hastables with different min/max values for either voltages or RPM, making some of the Hastbales currently useless. This is also a future fix that is planned to occur shortly. Unclear if it will be prepared in time for PNR on 11/2-3
changing hashtable size

Small changes based on meeting w/Harleen. Bringing plMode into the struct for display on dash eventually, and a bug w/ the PID math was also caught. Lastly, CAN was updated. We are running out of room in the 3rd message, so some alterations may need to happen. Also, giving a comment header to powerLimit.c & .h

Added catch all to LUT derived TQ setpoints incase we OOB somehow. Also, uploading the code I used to prove my pointers for the LUTs work. As a final comment, I will remove the plMode shortly because it is a waste and doesn't need to be had. it specifically exists for dash and 1 use in the POWERLIMIT_calculateTorqueCommand, and ultimately feels like a waste. I'm perfectly fine with recalculating which Mode we are in in any sort of getter function bc of the low use rate. In like 5 years this might be a small dash-only bug if they make changes to the avaliable PL mapping but not the getter function, but tbh thats on the individuals editing the code if they don't understand what they are looking at.

This commit should prevent a bug with the difference between dNm and Nm

Moving plMode logic to proper location. Additionally, further prep for a rotary dial on the steering wheel

Final sudocode setup for projected steering button inputs

corrected a  pointer in POWERLIMIT_populateHashTable to be set with the correct LUT